requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
dnspython==2.7.0
eval_type_backport==0.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jsonpath-python==1.0.6
mistralai==1.2.5
mypy-extensions==1.0.0
numpy==2.1.3
pandas==2.2.3
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
